homework11/task1/meta_class/meta_class.py
"""
Vasya implemented nonoptimal Enum classes.
Remove duplications in variables declarations using metaclasses.
from enum import Enum
class ColorsEnum(Enum):
    RED = "RED"
    BLUE = "BLUE"
    ORANGE = "ORANGE"
    BLACK = "BLACK"
class SizesEnum(Enum):
    XL = "XL"
    L = "L"
    M = "M"
    S = "S"
    XS = "XS"
Should become:
class ColorsEnum(metaclass=SimplifiedEnum):
    __keys = ("RED", "BLUE", "ORANGE", "BLACK")
class SizesEnum(metaclass=SimplifiedEnum):
    __keys = ("XL", "L", "M", "S", "XS")
assert ColorsEnum.RED == "RED"
assert SizesEnum.XL == "XL"
"""


class SimplifiedEnum(type):
    def __new__(cls, name, bases, attrs):
        cls_instance = super().__new__(cls, name, bases, attrs)
        setattr(cls_instance, "name_attr", f"_{name}__keys")
        return cls_instance

    def __getattr__(self, key):
        if key in self.__dict__[self.name_attr]:
            return key
        raise KeyError(f'The key - "{key}" does not exist')
